# ENH: Add SIMD sin/cos implementation with numpy-simd-routines