--- a/.gitignore
+++ b/.gitignore
@@ -50,6 +50,7 @@ build
 build-*
 # meson python output
 .mesonpy-native-file.ini
+subprojects/.wraplock
 # sphinx build directory
 _build
 # dist directory is where sdist/wheel end up